Module Re

Module Re: code for creating and using regular expressions, independently of regular expression syntax.

type t

Regular expression

type re

Compiled regular expression

module Group : sig ... end

Manipulate matching groups.

type groups = Group.t

deprecated Use Group.t

Compilation and execution of a regular expression

val compile : t -> re @@ portable

Compile a regular expression into an executable version that can be used to match strings, e.g. with exec.

val group_count : re -> int @@ portable

Return the number of capture groups (including the one corresponding to the entire regexp).

val group_names : re -> (string * int) list @@ portable

Return named capture groups with their index.

val exec : ?pos:int -> ?len:int -> re -> string -> Group.t @@ portable

exec re str searches str for a match of the compiled expression re, and returns the matched groups if any.

More specifically, when a match exists, exec returns a match that starts at the earliest position possible. If multiple such matches are possible, the one specified by the match semantics described below is returned.

Examples:

  # let regex = Re.compile Re.(seq [str "//"; rep print ]);;
  val regex : re = <abstr>

  # Re.exec regex "// a C comment";;
  - : Re.Group.t = <abstr>

  # Re.exec regex "# a C comment?";;
  Exception: Not_found

  # Re.exec ~pos:1 regex "// a C comment";;
  Exception: Not_found

parameter pos
optional beginning of the string (default 0)

parameter len
length of the substring of str that can be matched (default -1, meaning to the end of the string)

raises Not_found
if the regular expression can't be found in str

val exec_opt : ?pos:int -> ?len:int -> re -> string -> Group.t option @@ portable

Similar to exec, but returns an option instead of using an exception.

Examples:

  # let regex = Re.compile Re.(seq [str "//"; rep print ]);;
  val regex : re = <abstr>

  # Re.exec_opt regex "// a C comment";;
  - : Re.Group.t option = Some <abstr>

  # Re.exec_opt regex "# a C comment?";;
  - : Re.Group.t option = None

  # Re.exec_opt ~pos:1 regex "// a C comment";;
  - : Re.Group.t option = None

val execp : ?pos:int -> ?len:int -> re -> string -> bool @@ portable

Similar to exec, but returns true if the expression matches, and false if it doesn't. This function is more efficient than calling exec or exec_opt and ignoring the returned group.

Examples:

  # let regex = Re.compile Re.(seq [str "//"; rep print ]);;
  val regex : re = <abstr>

  # Re.execp regex "// a C comment";;
  - : bool = true

  # Re.execp ~pos:1 regex "// a C comment";;
  - : bool = false

val exec_partial : ?pos:int -> ?len:int -> re -> string -> [ `Full | `Partial | `Mismatch ] @@ portable

More detailed version of execp. `Full is equivalent to true, while `Mismatch and `Partial are equivalent to false, but `Partial indicates the input string could be extended to create a match.

Examples:

  # let regex = Re.compile Re.(seq [bos; str "// a C comment"]);;
  val regex : re = <abstr>

  # Re.exec_partial regex "// a C comment here.";;
  - : [ `Full | `Mismatch | `Partial ] = `Full

  # Re.exec_partial regex "// a C comment";;
  - : [ `Full | `Mismatch | `Partial ] = `Partial

  # Re.exec_partial regex "//";;
  - : [ `Full | `Mismatch | `Partial ] = `Partial

  # Re.exec_partial regex "# a C comment?";;
  - : [ `Full | `Mismatch | `Partial ] = `Mismatch

val exec_partial_detailed : ?pos:int -> ?len:int -> re -> string -> [ `Full of Group.t | `Partial of int | `Mismatch ] @@ portable

More detailed version of exec_opt. `Full group is equivalent to Some group, while `Mismatch and `Partial _ are equivalent to None, but `Partial position indicates that the input string could be extended to create a match, and no match could start in the input string before the given position. This could be used to not have to search the entirety of the input if more becomes available, and use the given position as the ?pos argument.

module Mark : sig ... end

Marks

High Level Operations

type split_token = [

| `Text of string
(*
Text between delimiters
*)

| `Delim of Group.t
(*
Delimiter
*)

]

val all : ?pos:int -> ?len:int -> re -> string -> Group.t list @@ portable

Repeatedly calls exec on the given string, starting at given position and length.

Examples:

  # let regex = Re.compile Re.(seq [str "my"; blank; word(rep alpha)]);;
  val regex : re = <abstr>

  # Re.all regex "my head, my shoulders, my knees, my toes ...";;
  - : Re.Group.t list = [<abstr>; <abstr>; <abstr>; <abstr>]

  # Re.all regex "My head, My shoulders, My knees, My toes ...";;
  - : Re.Group.t list = []

type 'a gen = unit -> 'a option

val all_gen : ?pos:int -> ?len:int -> re -> string -> Group.t gen @@ portable

deprecated
Use Seq.all instead.

val all_seq : ?pos:int -> ?len:int -> re -> string -> Group.t Seq.t @@ portable

deprecated
Use Seq.all instead.

val matches : ?pos:int -> ?len:int -> re -> string -> string list @@ portable

Same as all, but extracts the matched substring rather than returning the whole group. This basically iterates over matched strings.

Examples:

  # let regex = Re.compile Re.(seq [str "my"; blank; word(rep alpha)]);;
  val regex : re = <abstr>

  # Re.matches regex "my head, my shoulders, my knees, my toes ...";;
  - : string list = ["my head"; "my shoulders"; "my knees"; "my toes"]

  # Re.matches regex "My head, My shoulders, My knees, My toes ...";;
  - : string list = []

  # Re.matches regex "my my my my head my 1 toe my ...";;
  - : string list = ["my my"; "my my"]

  # Re.matches ~pos:2 regex "my my my my head my +1 toe my ...";;
  - : string list = ["my my"; "my head"]

val matches_gen : ?pos:int -> ?len:int -> re -> string -> string gen @@ portable

deprecated
Use Seq.matches instead.

val matches_seq : ?pos:int -> ?len:int -> re -> string -> string Seq.t @@ portable

deprecated
Use Seq.matches instead.

val split : ?pos:int -> ?len:int -> re -> string -> string list @@ portable

split re s splits s into chunks separated by re. It yields the chunks themselves, not the separator. An occurence of the separator at the beginning or the end of the string is ignoring.

Examples:

  # let regex = Re.compile (Re.char ',');;
  val regex : re = <abstr>

  # Re.split regex "Re,Ocaml,Jerome Vouillon";;
  - : string list = ["Re"; "Ocaml"; "Jerome Vouillon"]

  # Re.split regex "No commas in this sentence.";;
  - : string list = ["No commas in this sentence."]

  # Re.split regex ",1,2,";;
  - : string list = ["1"; "2"]

  # Re.split ~pos:3 regex "1,2,3,4. Commas go brrr.";;
  - : string list = ["3"; "4. Commas go brrr."]

Zero-length patterns:

Be careful when using split with zero-length patterns like eol, bow, and eow. Because they don't have any width, they will still be present in the result. (Note the position of the \n and space characters in the output.)

  # Re.split (Re.compile Re.eol) "a\nb";;
  - : string list = ["a"; "\nb"]

  # Re.split (Re.compile Re.bow) "a b";;
  - : string list = ["a "; "b"]

  # Re.split (Re.compile Re.eow) "a b";;
  - : string list = ["a"; " b"]

Compare this to the behavior of splitting on the char itself. (Note that the delimiters are not present in the output.)

  # Re.split (Re.compile (Re.char '\n')) "a\nb";;
  - : string list = ["a"; "b"]

  # Re.split (Re.compile (Re.char ' ')) "a b";;
  - : string list = ["a"; "b"]

val split_delim : ?pos:int -> ?len:int -> re -> string -> string list @@ portable

split_delim re s splits s into chunks separated by re. It yields the chunks themselves, not the separator. Occurences of the separator at the beginning or the end of the string will produce empty chunks.

Examples:

  # let regex = Re.compile (Re.char ',');;
  val regex : re = <abstr>

  # Re.split regex "Re,Ocaml,Jerome Vouillon";;
  - : string list = ["Re"; "Ocaml"; "Jerome Vouillon"]

  # Re.split regex "No commas in this sentence.";;
  - : string list = ["No commas in this sentence."]

  # Re.split regex ",1,2,";;
  - : string list = [""; "1"; "2"; ""]

  # Re.split ~pos:3 regex "1,2,3,4. Commas go brrr.";;
  - : string list = ["3"; "4. Commas go brrr."]

Zero-length patterns:

Be careful when using split_delim with zero-length patterns like eol, bow, and eow. Because they don't have any width, they will still be present in the result. (Note the position of the \n and space characters in the output.)

  # Re.split_delim (Re.compile Re.eol) "a\nb";;
  - : string list = ["a"; "\nb"; ""]

  # Re.split_delim (Re.compile Re.bow) "a b";;
  - : string list = [""; "a "; "b"]

  # Re.split_delim (Re.compile Re.eow) "a b";;
  - : string list = ["a"; " b"; ""]

Compare this to the behavior of splitting on the char itself. (Note that the delimiters are not present in the output.)

  # Re.split_delim (Re.compile (Re.char '\n')) "a\nb";;
  - : string list = ["a"; "b"]

  # Re.split_delim (Re.compile (Re.char ' ')) "a b";;
  - : string list = ["a"; "b"]

val split_gen : ?pos:int -> ?len:int -> re -> string -> string gen @@ portable

deprecated
Use Seq.split instead.

val split_seq : ?pos:int -> ?len:int -> re -> string -> string Seq.t @@ portable

deprecated
Use Seq.split instead.

val split_full : ?pos:int -> ?len:int -> re -> string -> split_token list @@ portable

split re s splits s into chunks separated by re. It yields the chunks along with the separators. For instance this can be used with a whitespace-matching re such as "[\t ]+".

Examples:

  # let regex = Re.compile (Re.char ',');;
  val regex : re = <abstr>

  # Re.split_full regex "Re,Ocaml,Jerome Vouillon";;
  - : Re.split_token list =
    [`Text "Re"; `Delim <abstr>; `Text "Ocaml"; `Delim <abstr>;
    `Text "Jerome Vouillon"]

  # Re.split_full regex "No commas in this sentence.";;
  - : Re.split_token list = [`Text "No commas in this sentence."]

  # Re.split_full ~pos:3 regex "1,2,3,4. Commas go brrr.";;
  - : Re.split_token list =
    [`Delim <abstr>; `Text "3"; `Delim <abstr>; `Text "4. Commas go brrr."]

val split_full_gen : ?pos:int -> ?len:int -> re -> string -> split_token gen @@ portable

deprecated
Use Seq.split_full instead.

val split_full_seq : ?pos:int -> ?len:int -> re -> string -> split_token Seq.t @@ portable

deprecated
Use Seq.split_full instead.

module Seq : sig ... end

String expressions (literal match)

val str : string -> t @@ portable

val char : char -> t @@ portable

Basic operations on regular expressions

val alt : t list -> t @@ portable

Alternative.

alt [] is equivalent to empty.

By default, the leftmost match is preferred (see match semantics below).

val seq : t list -> t @@ portable

Sequence

val empty : t @@ portable

Match nothing

val epsilon : t @@ portable

Empty word

val rep : t -> t @@ portable

0 or more matches

val rep1 : t -> t @@ portable

1 or more matches

val repn : t -> int -> int option -> t @@ portable

repn re i j matches re at least i times and at most j times, bounds included. j = None means no upper bound.

val opt : t -> t @@ portable

0 or 1 matches

String, line, word

We define a word as a sequence of latin1 letters, digits and underscore.

val bol : t @@ portable

Beginning of line

val eol : t @@ portable

End of line

val bow : t @@ portable

Beginning of word

val eow : t @@ portable

End of word

val bos : t @@ portable

Beginning of string. This differs from start because it matches the beginning of the input string even when using ~pos arguments:

  let b = execp (compile (seq [ bos; str "a" ])) "aa" ~pos:1 in
  assert (not b)

val eos : t @@ portable

End of string. This is different from stop in the way described in bos.

val leol : t @@ portable

Last end of line or end of string

val start : t @@ portable

Initial position. This differs from bos because it takes into account the ~pos arguments:

  let b = execp (compile (seq [ start; str "a" ])) "aa" ~pos:1 in
  assert b

val stop : t @@ portable

Final position. This is different from eos in the way described in start.

val word : t -> t @@ portable

Word

val not_boundary : t @@ portable

Not at a word boundary

val whole_string : t -> t @@ portable

Only matches the whole string, i.e. fun t -> seq [ bos; t; eos ].

Match semantics

A regular expression frequently matches a string in multiple ways. For instance exec (compile (opt (str "a"))) "ab" can match "" or "a". Match semantic can be modified with the functions below, allowing one to choose which of these is preferable.

By default, the leftmost branch of alternations is preferred, and repetitions are greedy.

Note that the existence of matches cannot be changed by specifying match semantics. seq [ bos; str "a"; non_greedy (opt (str "b")); eos ] will match when applied to "ab". However if seq [ bos; str "a"; non_greedy (opt (str "b")) ] is applied to "ab", it will match "a" rather than "ab".

Also note that multiple match semantics can conflict. In this case, the one executed earlier takes precedence. For instance, any match of shortest (seq [ bos; group (rep (str "a")); group (rep (str "a")); eos ]) will always have an empty first group. Conversely, if we use longest instead of shortest, the second group will always be empty.

val longest : t -> t @@ portable

Longest match semantics. That is, matches will match as many bytes as possible. If multiple choices match the maximum amount of bytes, the one respecting the inner match semantics is preferred.

val shortest : t -> t @@ portable

Same as longest, but matching the least number of bytes.

val first : t -> t @@ portable

First match semantics for alternations (not repetitions). That is, matches will prefer the leftmost branch of the alternation that matches the text.

val greedy : t -> t @@ portable

Greedy matches for repetitions (opt, rep, rep1, repn): they will match as many times as possible.

val non_greedy : t -> t @@ portable

Non-greedy matches for repetitions (opt, rep, rep1, repn): they will match as few times as possible.

Groups (or submatches)

val group : ?name:string -> t -> t @@ portable

Delimit a group. The group is considered as matching if it is used at least once (it may be used multiple times if is nested inside rep for instance). If it is used multiple times, the last match is what gets captured.

val no_group : t -> t @@ portable

Remove all groups

val nest : t -> t @@ portable

When matching against nest e, only the group matching in the last match of e will be considered as matching.

For instance:

  let re = compile (rep1 (nest (alt [ group (str "a"); str "b" ]))) in
  let group = Re.exec re "ab" in
  assert (Group.get_opt group 1 = None);
  (* same thing but without [nest] *)
  let re = compile (rep1 (alt [ group (str "a"); str "b" ])) in
  let group = Re.exec re "ab" in
  assert (Group.get_opt group 1 = Some "a")

val mark : t -> Mark.t * t @@ portable

Mark a regexp. the markid can then be used to know if this regexp was used.

Character sets

val set : string -> t @@ portable

Any character of the string

val rg : char -> char -> t @@ portable

Character ranges

val inter : t list -> t @@ portable

Intersection of character sets

val diff : t -> t -> t @@ portable

Difference of character sets

val compl : t list -> t @@ portable

Complement of union

Predefined character sets

val any : t @@ portable

Any character

val notnl : t @@ portable

Any character but a newline

val alnum : t @@ portable

val wordc : t @@ portable

val alpha : t @@ portable

val ascii : t @@ portable

val blank : t @@ portable

val cntrl : t @@ portable

val digit : t @@ portable

val graph : t @@ portable

val lower : t @@ portable

val print : t @@ portable

val punct : t @@ portable

val space : t @@ portable

val upper : t @@ portable

val xdigit : t @@ portable

Case modifiers

val case : t -> t @@ portable

Case sensitive matching. Note that this works on latin1, not ascii and not utf8.

val no_case : t -> t @@ portable

Case insensitive matching. Note that this works on latin1, not ascii and not utf8.

Internal debugging

val pp : Format.formatter -> t -> unit @@ portable

val pp_re : Format.formatter -> re -> unit @@ portable

val print_re : Format.formatter -> re -> unit @@ portable

Alias for pp_re. Deprecated

Experimental functions

val witness : t -> string @@ portable

witness r generates a string s such that execp (compile r) s is true.

Be warned that this function is buggy because it ignores zero-width assertions like beginning of words. As a result it can generate incorrect results.

Deprecated functions

type substrings = Group.t

Alias for Group.t. Deprecated

deprecated Use Group.t

val get : Group.t -> int -> string @@ portable

Same as Group.get. Deprecated

deprecated Use Group.get

val get_ofs : Group.t -> int -> int * int @@ portable

Same as Group.offset. Deprecated

deprecated Use Group.offset

val get_all : Group.t -> string array @@ portable

Same as Group.all. Deprecated

deprecated Use Group.all

val get_all_ofs : Group.t -> (int * int) array @@ portable

Same as Group.all_offset. Deprecated

deprecated Use Group.all_offset

val test : Group.t -> int -> bool @@ portable

Same as Group.test. Deprecated

deprecated Use Group.test

type markid = Mark.t

Alias for Mark.t. Deprecated

deprecated Use Mark.

val marked : Group.t -> Mark.t -> bool @@ portable

Same as Mark.test. Deprecated

deprecated Use Mark.test

val mark_set : Group.t -> Mark.Set.t @@ portable

Same as Mark.all. Deprecated

deprecated Use Mark.all

module Stream : sig ... end

An experimental for matching a regular expression by feeding individual string chunks.

val replace : ?pos:int -> ?len:int -> ?all:bool -> Re__.Compile.re -> f:(Group.t -> string) -> string -> string @@ portable

replace ~all re ~f s iterates on s, and replaces every occurrence of re with f substring where substring is the current match. If all = false, then only the first occurrence of re is replaced.

val replace_string : ?pos:int -> ?len:int -> ?all:bool -> Re__.Compile.re -> by:string -> string -> string @@ portable

replace_string ~all re ~by s iterates on s, and replaces every occurrence of re with by. If all = false, then only the first occurrence of re is replaced.

Examples:

  # let regex = Re.compile (Re.char ',');;
  val regex : re = <abstr>

  # Re.replace_string regex ~by:";" "[1,2,3,4,5,6,7]";;
  - : string = "[1;2;3;4;5;6;7]"

  # Re.replace_string regex ~all:false ~by:";" "[1,2,3,4,5,6,7]";;
  - : string = "[1;2,3,4,5,6,7]"

module View : sig ... end

A view of the top-level of a regex. This type is unstable and may change

module Emacs : sig ... end

Emacs-style regular expressions

module Glob : sig ... end

Shell-style regular expressions

module Perl : sig ... end

Perl-style regular expressions

module Pcre : sig ... end

NOTE: Only a subset of the PCRE spec is supported

module Posix : sig ... end

References:

module Str : sig ... end

Module Str: regular expressions and high-level string processing