Module Re.Pcre

exception Parse_error

exception Not_supported

type regexp

type flag = [

| `CASELESS

| `MULTILINE

| `ANCHORED

| `DOTALL

]

type groups = Group.t

type split_result =

| Text of string
(*
Text part of splitted string
*)

| Delim of string
(*
Delimiter part of splitted string
*)

| Group of int * string
(*
Subgroup of matched delimiter (subgroup_nr, subgroup_str)
*)

| NoGroup
(*
Unmatched subgroup
*)

Result of a Pcre.full_split

val re : ?flags:flag list -> string -> (Re__.Ast.cset, [ `Cased | `Uncased ]) Re__.Ast.gen

re ~flags s creates the regexp s using the pcre syntax.

val regexp : ?flags:flag list -> string -> regexp

re ~flags s compiles the regexp s using the pcre syntax.

val extract : rex:regexp -> string -> string array

extract ~rex s executes rex on s and returns the matching groups.

val exec : rex:regexp -> ?pos:int -> string -> groups

Equivalent to Core.exec.

val get_substring : groups -> int -> string

Equivalent to Core.Group.get.

val names : regexp -> string array

Return the names of named groups.

val get_named_substring : regexp -> string -> groups -> string

Return the first matched named group, or raise Not_found.

val get_substring_ofs : groups -> int -> int * int

Equivalent to Core.Group.offset.

val pmatch : rex:regexp -> string -> bool

Equivalent to Core.execp.

val substitute : rex:Re__.Compile.re -> subst:(string -> string) -> string -> string

val full_split : ?max:int -> rex:regexp -> string -> split_result list

val split : rex:regexp -> string -> string list

val quote : string -> string

Deprecated

type substrings = Group.t