Module Re.Stream

An experimental for matching a regular expression by feeding individual string chunks.

This module is not covered by semver's stability guarantee.

type t

type 'a feed =

| Ok of 'a

| No_match

val create : Re__.Compile.re -> t

val feed : t -> string -> pos:int -> len:int -> t feed

val finalize : t -> string -> pos:int -> len:int -> bool

finalize s ~pos ~len feed s from pos to len and return whether the regular expression matched.

module Group : sig ... end

Match a string against a regular expression with capture groups